SANC-44: Update booking form + move calendar view component location
